+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [17.17.0]
+- Added `DateTimeView` component
+- Made `DateView` and `SelectableDateViewModel` part of the public API
+
 ## [17.16.0]
 - [FilledCheckBox][SaveView][ContentSavePage] Changed the way the filled checkedbox looks when its not checked.